Create child classes for apb_push and apb_relist
